+++ src-tauri/src/filesystem/mod.rs
@@ -1,3 +1,4 @@
+pub mod actions;
 pub mod archive;
 pub mod fs;
 pub mod sftp;